PCINTS-325: Fixes fir .NET template.
Set the Prefer32Bit property to false in the project file.
Corrected the locations of the source files in the project file ( use '\' not '/' ).
Add a startup script to the blade.
Set the default config file to TemplatePricingAdapter.conf in code.
Rename the blade config file to TemplatePricingAdapter.conf.

Note that when I built this on Windows the macro reading worked.
When we build the kit it doesn't copy all the libraries to bin/Release just the .NET one.